profileにgroup_idのインデックスが存在しなければ、インデックス削除をスキップするよう実装

+++ db/migrate/20240927031307_remove_index_on_group_id_from_profiles.rb
@@ -1,5 +1,7 @@
 class RemoveIndexOnGroupIdFromProfiles < ActiveRecord::Migration[7.1]
-  def change
+  if index_exists?(:profiles, :group_id)
     remove_index :profiles, :group_id
+  else
+    puts "Index does not exist, skipping migration."
   end
 end
